Separate fastify app & server to testing easy
