+++ order-app/src/index.ts
@@ -1,44 +1,11 @@
 import dotenv from 'dotenv';
-import fastify from 'fastify';
-import { FastifyCorsOptions } from 'fastify-cors';
+import build from './app';
 
 dotenv.config();
 
-const server = fastify();
+const app = build({});
 
-const corsConfig = (): FastifyCorsOptions => ({
-  origin: (origin, cb) => {
-    // Allow is env development
-    if (process.env.NODE_ENV === 'development') {
-      cb(null, true);
-      return;
-    }
-
-    // Allow localhost
-    if (/localhost/.test(origin)) {
-      cb(null, true);
-      return;
-    }
-
-    cb(new Error('Not allowed'), false);
-  },
-});
-
-const databaseConfig = () => ({
-  client: 'pg',
-  connection: process.env.POSTGRESQL_CONNECTION_STRING,
-  pool: { min: 0, max: 2 },
-  searchPath: ['knex', 'public'],
-});
-
-// server.register(require('fastify-knexjs'), databaseConfig);
-server.register(require('./plugins/knex'), databaseConfig);
-server.register(require('fastify-cors'), corsConfig);
-server.register(require('./routes/orders'), { prefix: '/orders' });
-
-server.get('/ping', async () => 'pong pong pong\n');
-
-server.listen(4000, (err, address) => {
+app.listen(4000, (err, address) => {
   if (err) {
     console.error(err);
     process.exit(1);
